requirements.txt
Django==2.2
django-extensions==2.2.5
djangorestframework==3.10.3
factory-boy==2.12.0
ipython==7.9.0
psycopg2==2.8.4